chore: Update dataset URL to version 2.01.1
The code changes update the dataset URL in the `gdp1h` function to version 2.01.1. This ensures that the latest version of the dataset is being used. The previous URL was "https://noaa-oar-hourly-gdp-pds.s3.amazonaws.com/latest/gdp-v2.01.zarr" and it has been updated to "https://noaa-oar-hourly-gdp-pds.s3.amazonaws.com/latest/gdp-v2.01.1.zarr".

Co-authored-by: Shane Elipot <selipot@miami.edu>
Co-authored-by: Philippe Miron <philippemiron@gmail.com>